+++ controller/src/main/java/com/linbit/linstor/core/apicallhandler/controller/internal/CtrlSatelliteUpdateCaller.java
@@ -550,7 +550,7 @@ public class CtrlSatelliteUpdateCaller
         try
         {
             return updateSatellites(
-                atomicUpdateDataRef.getInvolvedNodes(apiCtx),
+                atomicUpdateDataRef.getInvolvedOnlineNodes(apiCtx),
                 atomicUpdateDataRef,
                 notConnectedErrorRef
             );